+++ PLC_esp8266/main/LogicProgram/Inputs/TimerMSecs.cpp
@@ -30,7 +30,7 @@ void TimerMSecs::SetTime(uint32_t delay_time_ms) {
         delay_time_ms = TimerMSecs::max_delay_time_ms;
     }
     this->delay_time_us = delay_time_ms * 1000LL;
-    str_size = sprintf(this->str_time, "%u", delay_time_ms);
+    str_size = sprintf(this->str_time, "%u", (unsigned int)delay_time_ms);
 
     ESP_LOGD(TAG_TimerMSecs, "ctor, str_time:%s", this->str_time);
 }
